# io/rest/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    ParkingAdminCreate:
        properties:
            enabled:
                type: boolean
                x-go-name: Enabled
            first_name:
                type: string
                x-go-name: FirstName
            id:
                format: int64
                type: integer
                x-go-name: ID
            last_name:
                type: string
                x-go-name: LastName
            parking_id:
                format: int64
                type: integer
                x-go-name: ParkingID
            phone:
                type: string
                x-go-name: Phone
        type: object
        x-go-package: github.com/farazff/IoT-parking/io/rest/sw
    ParkingAdminCreator:
        properties:
            enabled:
                type: boolean
                x-go-name: Enabled
            first_name:
                type: string
                x-go-name: FirstName
            last_name:
                type: string
                x-go-name: LastName
            parking_id:
                format: int64
                type: integer
                x-go-name: ParkingID
            password:
                type: string
                x-go-name: Password
            phone:
                type: string
                x-go-name: Phone
        required:
            - first_name
            - last_name
            - phone
            - enabled
            - password
            - parking_id
        type: object
        x-go-package: github.com/farazff/IoT-parking/io/rest/sw
    ParkingAdminGet:
        properties:
            enabled:
                type: boolean
                x-go-name: Enabled
            first_name:
                type: string
                x-go-name: FirstName
            id:
                format: int64
                type: integer
                x-go-name: ID
            last_name:
                type: string
                x-go-name: LastName
            parking_id:
                format: int64
                type: integer
                x-go-name: ParkingID
            phone:
                type: string
                x-go-name: Phone
        type: object
        x-go-package: github.com/farazff/IoT-parking/io/rest/sw
    ParkingAdminUpdater:
        properties:
            enabled:
                type: boolean
                x-go-name: Enabled
            first_name:
                type: string
                x-go-name: FirstName
            last_name:
                type: string
                x-go-name: LastName
            parking_id:
                format: int64
                type: integer
                x-go-name: ParkingID
            phone:
                type: string
                x-go-name: Phone
        required:
            - first_name
            - last_name
            - phone
            - enabled
            - parking_id
        type: object
        x-go-package: github.com/farazff/IoT-parking/io/rest/sw
...
    ZoneCreator:
        properties:
            capacity:
                type: string
                x-go-name: Capacity
            enabled:
                type: string
                x-go-name: Enabled
            remained_capacity:
                type: string
                x-go-name: RemainedCapacity
        required:
            - capacity
            - enabled
            - remained_capacity
        type: object
        x-go-package: github.com/farazff/IoT-parking/io/rest/sw
    ZoneGet:
        properties:
            enabled:
                type: boolean
                x-go-name: Enabled
            first_name:
                type: string
                x-go-name: FirstName
            id:
                format: int64
                type: integer
                x-go-name: ID
            last_name:
                type: string
                x-go-name: LastName
            parking_id:
                format: int64
                type: integer
                x-go-name: ParkingID
            phone:
                type: string
                x-go-name: Phone
        type: object
        x-go-package: github.com/farazff/IoT-parking/io/rest/sw
...
host: localhost:7676
info:
    contact:
        email: f.farangizadeh@gmail.com
        name: Faraz Farangizadeh
    description: Document of Parking-IoT api
    title: Parking IoT Swagger
    version: 1.0.0
paths:
...
    /v1/parkingAdmin:
        post:
            operationId: createParkingAdmin
            parameters:
                - description: api key for APIs
                  in: header
                  name: api-key
                  required: true
                  type: string
                  x-go-name: ApiKey
                - description: Session token
                  in: header
                  name: session_token
                  required: true
                  type: string
                  x-go-name: SessionToken
                - in: body
                  name: ParkingAdminCreator
                  required: true
                  schema:
                    $ref: '#/definitions/ParkingAdminCreator'
            responses:
                "201":
                    $ref: '#/responses/ParkingAdminCreateRes'
                "400":
                    $ref: '#/responses/ErrorMessage'
                "401":
                    $ref: '#/responses/ErrorUnauthorizedMessage'
                "500":
                    $ref: '#/responses/ErrorMessage'
            summary: This route is used to create parking admin
            tags:
                - System_Admin
    /v1/parkingAdmin/{id}:
        delete:
            operationId: deleteParkingAdmin
            parameters:
                - description: api key for APIs
                  in: header
                  name: api-key
                  required: true
                  type: string
                  x-go-name: ApiKey
                - description: Session token
                  in: header
                  name: session_token
                  required: true
                  type: string
                  x-go-name: SessionToken
                - format: int64
                  in: path
                  name: ID
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/ErrorMessage'
                "400":
                    $ref: '#/responses/ErrorMessage'
                "401":
                    $ref: '#/responses/ErrorUnauthorizedMessage'
                "404":
                    $ref: '#/responses/ErrorMessage'
                "500":
                    $ref: '#/responses/ErrorMessage'
            summary: This route is used to delete a parking admin by ID
            tags:
                - System_Admin
        get:
            operationId: getParkingAdmin
            parameters:
                - description: api key for APIs
                  in: header
                  name: api-key
                  required: true
                  type: string
                  x-go-name: ApiKey
                - description: Session token
                  in: header
                  name: session_token
                  required: true
                  type: string
                  x-go-name: SessionToken
                - format: int64
                  in: path
                  name: ID
                  required: true
                  type: integer
            responses:
                "200":
                    $ref: '#/responses/ParkingAdminGetRes'
                "400":
                    $ref: '#/responses/ErrorMessage'
                "401":
                    $ref: '#/responses/ErrorUnauthorizedMessage'
                "404":
                    $ref: '#/responses/ErrorMessage'
                "500":
                    $ref: '#/responses/ErrorMessage'
            summary: This route is used to get a single parking admin by ID
            tags:
                - System_Admin
        put:
            operationId: updateParkingAdmin
            parameters:
                - in: body
                  name: parkingUpdater
                  required: true
                  schema:
                    $ref: '#/definitions/ParkingAdminUpdater'
                  x-go-name: ParkingAdminUpdater
                - description: api key for APIs
                  in: header
                  name: api-key
                  required: true
                  type: string
                  x-go-name: ApiKey
                - description: Session token
                  in: header
                  name: session_token
                  required: true
                  type: string
                  x-go-name: SessionToken
                - format: int64
                  in: path
                  name: ID
                  required: true
                  type: integer
                - in: body
                  name: zoneUpdater
                  required: true
                  schema:
                    $ref: '#/definitions/ZoneCreator'
                  x-go-name: ZoneCreator
            responses:
                "201":
                    $ref: '#/responses/ParkingAdminUpdateRes'
                "400":
                    $ref: '#/responses/ErrorMessage'
                "401":
                    $ref: '#/responses/ErrorUnauthorizedMessage'
                "404":
                    $ref: '#/responses/ErrorMessage'
                "500":
                    $ref: '#/responses/ErrorMessage'
            summary: This route is used to update a parking admin
            tags:
                - System_Admin
...
    /v1/parkingAdmins:
        get:
            operationId: getParkingAdmins
            parameters:
                - description: api key for APIs
                  in: header
                  name: api-key
                  required: true
                  type: string
                  x-go-name: ApiKey
                - description: Session token
                  in: header
                  name: session_token
                  required: true
                  type: string
                  x-go-name: SessionToken
            responses:
                "200":
                    $ref: '#/responses/ParkingAdminsGetRes'
                "400":
                    $ref: '#/responses/ErrorMessage'
                "401":
                    $ref: '#/responses/ErrorUnauthorizedMessage'
                "500":
                    $ref: '#/responses/ErrorMessage'
            summary: This route is used to get all parking admins
            tags:
                - System_Admin
...
produces:
    - application/json
responses:
    ErrorMessage:
        description: ""
        headers:
            session_token:
                description: 'in: cookie'
                type: string
        schema:
            properties:
                message:
                    description: Error message
                    type: string
                    x-go-name: Message
            type: object
    ErrorUnauthorizedMessage:
        description: ""
        schema:
            properties:
                message:
                    description: Error message
                    type: string
                    x-go-name: Message
            type: object
...
    ParkingAdminCreateRes:
        description: ""
        headers:
            session_token:
                description: 'in: cookie'
                type: string
        schema:
            properties:
                parking_admin:
                    $ref: '#/definitions/ParkingAdminCreate'
            type: object
    ParkingAdminGetRes:
        description: ""
        headers:
            session_token:
                description: 'in: cookie'
                type: string
        schema:
            properties:
                zone:
                    $ref: '#/definitions/ZoneGet'
            type: object
    ParkingAdminUpdateRes:
        description: ""
        headers:
            session_token:
                description: 'in: cookie'
                type: string
        schema:
            properties:
                parking_admin:
                    $ref: '#/definitions/ParkingAdminGet'
            type: object
    ParkingAdminsGetRes:
        description: ""
        headers:
            session_token:
                description: 'in: cookie'
                type: string
        schema:
            properties:
                parking_admins:
                    description: parking res
                    items:
                        $ref: '#/definitions/ParkingAdminGet'
                    type: array
                    x-go-name: ParkingAdminGet
            type: object
...
schemes:
    - http
swagger: "2.0"